split template to partials
