+++ package.json
@@ -64,7 +64,7 @@
   },
   "generator": {
     "renderer": "react",
-    "generator": ">=0.50.0 <2.0.0",
+    "generator": ">=1.0.1 <2.0.0",
     "parameters": {
       "outFilename": {
         "description": "The name of the output markdown file",
