+++ partials/Schema.js
@@ -1,0 +1,133 @@
+import React from "react";
+import { Text } from "@asyncapi/generator-react-sdk";
+
+import { Header, TableHead, TableRow } from "../partials/common";
+
+export function Schema({ schema, schemaName, hideTitle = false }) {
+  const headers = ['Name', 'Type', 'Description', 'Accepted values'];
+  
+  let properties = schema.properties();
+  properties = Object.keys(properties).length ? Object.entries(properties).map(([propName, prop]) => (
+    <SchemaProp prop={prop} propName={propName} path='' required={isRequired(schema, propName)} />
+  )) : <SchemaProp prop={schema} propName={schemaName} path='' />;
+
+  return (
+    <Text>
+      {hideTitle === false ? <Header type={4}>{schemaName}</Header> : null}
+      <TableHead headers={headers} />
+      {properties}
+    </Text>
+  );
+}
+
+function SchemaProp({ prop, propName, required = false, path = '', circularPropsParent = [] }) {
+  const anyOf = prop.anyOf() && prop.anyOf().map((p, idx) => (
+    <SchemaProp prop={p} propName={idx} path={propName || buildPath(path, idx)} />
+  ));
+  const allOf = prop.allOf() && prop.allOf().map((p, idx) => (
+    <SchemaProp prop={p} propName={idx} path={propName || buildPath(path, idx)} />
+  ));
+  const oneOf = prop.oneOf() && prop.oneOf().map((p, idx) => (
+    <SchemaProp prop={p} propName={idx} path={propName || buildPath(path, idx)} />
+  ));
+  
+  const properties = Object.entries(prop.properties()).map(([pName, p]) => {
+    const circProps = p.circularProps();
+    const isPropCircular = circularPropsParent.includes(pName);
+    
+    if (isPropCircular === true) {
+      return (
+        <SchemaPropRow
+          prop={p} 
+          propName={pName}
+          path={propName || buildPath(path, pName)} 
+          required={isRequired(prop, pName)}
+          isCircular={isPropCircular}
+        />
+      );
+    } else {
+      return (
+        <SchemaProp 
+          prop={p} 
+          propName={pName} 
+          path={propName || buildPath(path, pName)} 
+          required={isRequired(prop, pName)}
+          circularPropsParent={circProps}
+        />
+      );
+    }
+  });
+
+  const additionalProperties = prop.additionalProperties() && prop.additionalProperties().properties 
+    ? Object.entries(prop.additionalProperties()).map(([pName, p]) => (
+      <SchemaProp prop={p} propName={pName} path={propName || buildPath(path, pName)} required={isRequired(prop.additionalProperties(), pName)} />
+    )) : null;
+  
+  const items = prop.items() && prop.items().properties
+    ? Object.entries(prop.items().properties()).map(([pName, p]) => {
+      const isCirc = p.isCircular();
+
+      if (isCirc === true) {
+        return (
+          <SchemaPropRow
+            prop={p} 
+            propName={pName}
+            path={propName || buildPath(path, pName)} 
+            required={isRequired(prop, pName)}
+            isCircular={isCirc}
+          />
+        );
+      } else {
+        return (
+          <SchemaProp 
+            prop={p} 
+            propName={pName} 
+            path={propName || buildPath(path, pName)} 
+            required={isRequired(prop.items(), pName)}
+          />
+        );
+      }
+    }) : null;
+
+  return (
+    <>
+      <SchemaPropRow prop={prop} propName={propName} required={required} path={path} />
+      {anyOf}
+      {allOf}
+      {oneOf}
+      {properties}
+      {additionalProperties}
+      {items}
+    </>
+  );
+}
+
+function SchemaPropRow({ prop, propName, required = false, path = '', isCircular = false }) {
+  const acceptedValues = prop.enum() && prop.enum().length ? prop.enum().join(', ') : '_Any_';
+  const types = `${prop.anyOf() ? `anyOf` : ''}${prop.allOf() ? `allOf` : ''}${prop.oneOf() ? `oneOf` : ''}${prop.items() && !Array.isArray(prop.items()) && prop.items().type() ? `${prop.items().type()}` : ''}`;
+  let description = `${prop.description() || ''}${isCircular ? ' **[CIRCULAR]**': ''}`
+  const rowRenderer = () => [
+    `${tree(path, propName)}${required ? ' **(required)**': ''}`,
+    `${prop.type() || ''}${types}`,
+    description.trim() || '-',
+    acceptedValues
+  ];
+
+  return <TableRow rowRenderer={rowRenderer} entry={prop} />;
+}
+
+function tree(path, field = '') {
+  const filteredPaths = path.split('.').filter(Boolean);
+  if (!filteredPaths.length) return field;
+  const dottedPath = filteredPaths.join('.');
+  return `${dottedPath}.${field}`;
+}
+
+function buildPath(propName, path) {
+  if (!path) return propName;
+  return `${path}.${propName}`;
+}
+
+function isRequired(obj, key) {
+  return obj && Array.isArray(obj.required) && !!(obj.required.includes(key));
+}
